Updated the configuration order and also tests for new workflow related chnages. Removed additional authorization for service level api calls for draft controller.

+++ java/customization-api/src/main/java/gov/nist/oar/customizationapi/web/DraftController.java
@@ -22,12 +22,15 @@ import org.bson.Document;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
+import org.springframework.security.authentication.InternalAuthenticationServiceException;
 //import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -45,23 +48,20 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 
 /**
- * This is a webservice/restapi controller which gives options to access, update
- * and delete the record. There are four end points provided in this, each
+ * This is a webservice/restapi controller which gives access to customization cache database.
+ * On behalf of MIDAS the metadata server can put data or record, delete or access it whenever needed.
+ * There are three end points provided in this, each
  * dealing with specific tasks. In OAR project internal landing page for the edi
  * record is accessed using backed metadata. This metadata is a advanced POD
- * record called NERDm. In this api we are allowing the record to be modified by
- * authorized user. This webservice connects to backend MongoDB which holds the
- * record being edited. When the record is accessed for the first time, it is
- * fetched from backend metadata service. If it gets modified the updated record
- * is saved in this stagging database until finalzed Once it is finalized it is
- * pushed back to backend service to merge and send to review.
+ * record called NERDm.  This webservice connects to backend MongoDB which holds the
+ * record being edited. The service needs an authorized token to access these endpoints.
  * 
  * @author Deoyani Nandrekar-Heinis
  *
  */
 @RestController
 @Api(value = "Api endpoints to access editable data, update changes to data, save in the backend", tags = "Customization API")
-@Validated
+//@Validated
 @RequestMapping("/pdr/lp/draft/")
 public class DraftController {
 	private Logger logger = LoggerFactory.getLogger(DraftController.class);
@@ -69,8 +69,12 @@ public class DraftController {
 	@Autowired
 	private DraftService draftRepo;
 
+	@Value("${custom.service.secret:testtoken}")
+	String authorization;
+
 	/***
-	 * Get complete record or only the changes made to the record by providing 'view=updates' option.
+	 * Get complete record or only the changes made to the record by providing
+	 * 'view=updates' option.
 	 * 
 	 * @param ediid Unique record identifier
 	 * @return Document
@@ -78,14 +82,16 @@ public class DraftController {
 	 */
 	@RequestMapping(value = { "{ediid}" }, method = RequestMethod.GET, produces = "application/json")
 	@ApiOperation(value = ".", nickname = "Access editable Record", notes = "Resource returns a record if it is editable and user is authenticated.")
-	public Document getData(@PathVariable @Valid String ediid, @RequestParam(required = false) String view) throws CustomizationException {
+	public Document getData(@PathVariable @Valid String ediid, @RequestParam(required = false) String view,
+			@RequestHeader("Authorization") String serviceAuth) throws CustomizationException {
 		logger.info("Access the record to be edited by ediid " + ediid);
+		if (!serviceAuth.equals(authorization))
+			throw new InternalAuthenticationServiceException("Service is not authorized to access this record.");
 		String viewoption = "";
-		if(view != null && !view.equals(""))
+		if (view != null && !view.equals(""))
 			viewoption = view;
-		return draftRepo.getDraft(ediid,viewoption);
+		return draftRepo.getDraft(ediid, viewoption);
 	}
-
 
 	/**
 	 * Delete the resource from staging area/cache
@@ -96,14 +102,18 @@ public class DraftController {
 	 */
 	@RequestMapping(value = { "{ediid}" }, method = RequestMethod.DELETE, produces = "application/json")
 	@ApiOperation(value = ".", nickname = "Delete the Record from drafts", notes = "This will allow user to delete all the changes made in the record in draft mode, original published record will remain as it is.")
-	public boolean deleteRecord(@PathVariable @Valid String ediid) throws CustomizationException {
+	public boolean deleteRecord(@PathVariable @Valid String ediid, @RequestHeader("Authorization") String serviceAuth)
+			throws CustomizationException {
 		logger.info("Delete the record from stagging given by ediid " + ediid);
+		if (!serviceAuth.equals(authorization))
+			throw new InternalAuthenticationServiceException("Service is not authorized to access this record.");
+
 		return draftRepo.deleteDraft(ediid);
 	}
 
 	/**
-	 * Finalize changes made in the record and send it back to backend metadata
-	 * server to merge and send for review.
+	 * Metadata server send data over to store in the cache/staging area until editing is done and finalized by 
+	 * client application.
 	 * 
 	 * @param ediid  Unique record id
 	 * @param params Modified fields in JSON
@@ -115,15 +125,18 @@ public class DraftController {
 			"{ediid}" }, method = RequestMethod.PUT, headers = "accept=application/json", produces = "application/json")
 	@ApiOperation(value = ".", nickname = "Save changes to server", notes = "Resource returns a boolean based on success or failure of the request.")
 	@ResponseStatus(HttpStatus.CREATED)
-	public void createRecord(@PathVariable @Valid String ediid, @Valid @RequestBody Document params)
+	public void createRecord(@PathVariable @Valid String ediid, @Valid @RequestBody Document params,
+			@RequestHeader("Authorization") String serviceAuth)
 			throws CustomizationException, InvalidInputException, ResourceNotFoundException {
 		logger.info("Send updated record to backend metadata server:" + ediid);
+		if (!serviceAuth.equals(authorization))
+			throw new InternalAuthenticationServiceException("Service is not authorized to access this record.");
 		draftRepo.putDraft(ediid, params);
 
 	}
 
 	/**
-	 * 
+	 * If there is an internal error due to certain functions failue this is called.
 	 * @param ex
 	 * @param req
 	 * @return
@@ -136,7 +149,7 @@ public class DraftController {
 	}
 
 	/**
-	 * 
+	 * If record is not available in the database
 	 * @param ex
 	 * @param req
 	 * @return
@@ -149,7 +162,7 @@ public class DraftController {
 	}
 
 	/**
-	 * 
+	 * If input is not of allowed format
 	 * @param ex
 	 * @param req
 	 * @return
@@ -162,7 +175,7 @@ public class DraftController {
 	}
 
 	/**
-	 * 
+	 * Some generic exception thrown by service
 	 * @param ex
 	 * @param req
 	 * @return
@@ -175,7 +188,7 @@ public class DraftController {
 	}
 
 	/**
-	 * 
+	 * If there is any runtime error
 	 * @param ex
 	 * @param req
 	 * @return
@@ -203,18 +216,18 @@ public class DraftController {
 		return new ErrorInfo(req.getRequestURI(), 502, "Can not connect to backend server");
 	}
 
-//	/**
-//	 * Handles internal authentication service exception if user is not authorized
-//	 * or token is expired
-//	 * 
-//	 * @param ex
-//	 * @param req
-//	 * @return
-//	 */
-//	@ExceptionHandler(InternalAuthenticationServiceException.class)
-//	@ResponseStatus(HttpStatus.UNAUTHORIZED)
-//	public ErrorInfo handleRestClientError(InternalAuthenticationServiceException ex, HttpServletRequest req) {
-//		logger.error("Unauthorized user or token : " + req.getRequestURI() + "\n  " + ex.getMessage(), ex);
-//		return new ErrorInfo(req.getRequestURI(), 401, "Untauthorized user or token.");
-//	}
+	/**
+	 * Handles internal authentication service exception if user is not authorized
+	 * or token is expired
+	 * 
+	 * @param ex
+	 * @param req
+	 * @return
+	 */
+	@ExceptionHandler(InternalAuthenticationServiceException.class)
+	@ResponseStatus(HttpStatus.UNAUTHORIZED)
+	public ErrorInfo handleRestClientError(InternalAuthenticationServiceException ex, HttpServletRequest req) {
+		logger.error("Unauthorized user or token : " + req.getRequestURI() + "\n  " + ex.getMessage(), ex);
+		return new ErrorInfo(req.getRequestURI(), 401, "Untauthorized token used to acces the service.");
+	}
 }
